requirements.txt
streamlit
torch
torchvision
fastai
opencv-python
imageio
Pillow